emailer/requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
google-cloud-functions>=1.0.0
protobuf>=3.20.0
grpcio>=1.50.0
grpcio-status>=1.50.0
pydantic>=2.0.0